requirements.txt
cdlib
networkx
pydantic
comet_ml
python-dotenv